+++ lib/domain/entities/enum_option_entity.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
 class EnumOptionEntity<T extends Enum> {
-  final T? type;
+  final T type;
   final IconData? icon;
   final String? label;
   final Color? iconColor;
 
-  EnumOptionEntity({this.type, this.icon, this.label, this.iconColor});
+  EnumOptionEntity({required this.type, this.icon, this.label, this.iconColor});
 }